+++ rrr-backend/src/main/java/com/nous/rollingrevenue/service/impl/StrategicBusinessUnitHeadServiceImpl.java
@@ -65,6 +65,16 @@ public class StrategicBusinessUnitHeadServiceImpl implements StrategicBusinessUn
 	@Override
 	@Transactional
 	public void deleteSBUHeadById(Long sbuHeadId) {
+		List<AnnualTargetEntry> annualTargetEntryList = annualTargetEntryRepository.findBySBUHeadId(sbuHeadId);
+		if (!annualTargetEntryList.isEmpty()) {
+			throw new RecordNotFoundException(
+					"SBU Head is already linked to AnnualTargetEntry or RevenueResourceEntry");
+		}
+		List<RevenueResourceEntry> revenueResourceList = revenueResourceEntryRepository.findBySBUHeadId(sbuHeadId);
+		if (!revenueResourceList.isEmpty()) {
+			throw new RecordNotFoundException(
+					"SBU Head is already linked to AnnualTargetEntry or RevenueResourceEntry");
+		}
 		sbuHeadRepository.findById(sbuHeadId)
 				.orElseThrow(() -> new RecordNotFoundException(ErrorConstants.RECORD_NOT_EXIST + sbuHeadId));
 		sbuHeadRepository.deleteById(sbuHeadId);
